@@ -13,12 +13,12 @@ using Microsoft.AspNetCore.Diagnostics;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.HttpsPolicy;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using Microsoft.Extensions.Hosting;
 using Microsoft.IdentityModel.Tokens;
 
 namespace DatingApp.API
@@ -37,7 +37,9 @@ namespace DatingApp.API
     {
       services.AddDbContext<DataContext>(options => options.UseSqlite(Configuration.GetConnectionString("DefaultConnection")));
 
-      services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2).AddJsonOptions(opt => opt.SerializerSettings.ReferenceLoopHandling = Newtonsoft.Json.ReferenceLoopHandling.Ignore);
+      services.AddRazorPages();
+      services.AddControllers()
+           .AddNewtonsoftJson(opt => opt.SerializerSettings.ReferenceLoopHandling = Newtonsoft.Json.ReferenceLoopHandling.Ignore);
 
       services.AddCors();
 
@@ -68,7 +70,7 @@ namespace DatingApp.API
     }
 
     // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-    public void Configure(IApplicationBuilder app, IHostingEnvironment env, Seed seeder)
+    public void Configure(IApplicationBuilder app, IWebHostEnvironment env, Seed seeder)
     {
       if (env.IsDevelopment())
       {
@@ -99,9 +101,22 @@ namespace DatingApp.API
 
       app.UseCors(x => x.AllowAnyOrigin().AllowAnyMethod().AllowAnyHeader());
 
+
       app.UseAuthentication();
 
-      app.UseMvc();
+      app.UseDefaultFiles();
+      app.UseStaticFiles();
+
+      app.UseRouting();
+      app.UseAuthorization();
+      app.UseEndpoints(endpoints =>
+          {
+            endpoints.MapControllerRoute(
+            name: "default",
+            pattern: "{controller=Home}/{action=Index}/{id?}");
+
+            endpoints.MapFallbackToController("fallback", "Index", "");
+          });
     }
   }
 }
